Rename 'plugin_hooks' to 'event_dispatch'
The old name was the name of two concepts the file was related to.
The new name is what process the header file actually controls
